requirements.txt
fastapi
fire
mypy
Pillow
